arptables: defeat multi-job for install
CQID: 399299

The make install target is not parallel job
safe, but doesn't do all that much, so we
limit make to one job for that.

The make target is of the form

install: x y z

Unfortunately, z depends on y, so if z is started
before y (almost) completes, we will fail.

Signed-off-by: Joe Slater <jslater@windriver.com>
Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>