Quagga: sync zebra routing table with the kernel one
When interface goes down, and there are some static routes that are
made with ip route, they are going to be deleted from kernel fib.
That's what kernel does by default. But they are not going to be
deleted from zebra's fib. They will be declared inactive.

This issue was originally discussed in [1] and a patch was proposed.

[1] http://www.gossamer-threads.com/lists/quagga/dev/22609

Signed-off-by: Aws Ismail <aws.ismail@windriver.com>
Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>