ntp: disable automatic panic shutdown for ntpd
Taken from [1]:

"However, if the difference between the client time and server time is greater
than the panic threshold, which defaults to 1000 s, the daemon will send a
message to the system log and shut down without setting the clock."

On RTC-less or 'fresh' systems with unpredicatble inital date ntpd will stop
with work undone (note: I did NOT find the message mentioned above - so it
took long way to fix). This is a wide use case for using NTP in production:
No need for initial datetime setup by some operator.

[1] http://doc.ntp.org/4.1.1/debug.htm

Signed-off-by: Andreas Müller <schnitzeltony@googlemail.com>
Signed-off-by: Joe MacDonald <joe@deserted.net>