ntpd: update config file comments
The default ntp.conf file does the right thing but one of the comments is
a bit misleading based on the active parts of the file changing a few
times.  Update the comment to accurately describe what is happening in the
configuration file and what a user should do next.

Signed-off-by: Joe MacDonald <joe@deserted.net>