crda: fix udev rule
The SBINDIR variable in makefile's install rule needs to have a trailing
slash so the path to the 'crda' binary in the udev rule is correct.

Otherwise the udev rule has a wrong path:

KERNEL=="regulatory*", ACTION=="change", SUBSYSTEM=="platform", RUN+="/usr/sbincrda"

and at runtime udev complains:

failed to execute '/usr/sbincrda'

when a regulatory domain event is triggered.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>
Acked-by: Koen Kooi <koen@dominion.thruhere.net>
Signed-off-by: Joe MacDonald <joe@deserted.net>